+++ src/constants/index.js
@@ -16,11 +16,23 @@ const IS_LINUX = (osName === 'linux');
 // Environment
 const IS_DEV = (env.name === 'development');
 
+// Settings
+const SETTING_TRAY_ENABLED = 'trayEnabledPref';
+const SETTING_TRAY_CLICK_SHORTCUT = 'trayClickShortcut';
+
+// Events
+const EVENT_WEBVIEW_NOTIFICATION = 'messages-webview-notification';
+const EVENT_NOTIFICATION_REFLECT_READY = 'messages-webview-reflect-ready';
+
 export {
     osName,
     osNameFriendly,
     IS_WINDOWS,
     IS_MAC,
     IS_LINUX,
-    IS_DEV
+    IS_DEV,
+    SETTING_TRAY_ENABLED,
+    SETTING_TRAY_CLICK_SHORTCUT,
+    EVENT_WEBVIEW_NOTIFICATION,
+    EVENT_NOTIFICATION_REFLECT_READY
 };